requirements.txt
future==1.0.0
iso8601==2.1.0
llvmlite==0.44.0
numba==0.61.0
numpy==2.1.3
pandas==2.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
serial==0.0.97
six==1.17.0
tzdata==2025.1
